+++ Professions/CHANGELOG.md
@@ -19,11 +19,12 @@
 * Skill reset feature is unchanged. But note that choosing to Master a skill will prevent it from being reset further, effectively locking you out of any unobtained profession. *This may change in the future, but for now you have been warned.*
 * Similarly, the Limit Break is now a reward for Mastering the Combat skill.
 * All prestiged profession variants, and a few base variants, have been reworked. Prestige professions are now much more impactful end-game rewards. Please review the section [Professions](README.md#professions) of the README to learn more.
+* Treasure Hunts are now triggered on Time Change instead of Player Warped. The chance to start a Treasure Hunt now depends on how far you have traveled or how many rocks you have broken since the previous hunt.
 * Misc. code improvements.
 
 ### Removed
 
-* Removed support for Luck skill and Love of Cooking. I might re-add Love of Cooking support if it gets updated.
+* Removed custom support for Luck skill and Love of Cooking.
 * Removed alost all third-party mod integrations since I don't know which of them were/will be updated. I might re-add Automate integration later, or I might not.